requirements.txt
fastapi==0.114.0
keras==3.5.0
numpy==2.1.1
Pillow==10.4.0
pydantic==2.9.1
